+++ src/i18n/Locale.ts
@@ -1,8 +1,6 @@
 import { durationFormatterForLocale } from './DurationFormatter';
 import type { PlayButton, TimeRange } from '../components';
 
-const localesByName: Record<string, Required<Locale>> = {};
-
 export interface Locale {
     /**
      * The {@link HTMLElement.ariaLabel | `aria-label`} for a {@link PlayButton} when it is showing a "play" button,
@@ -55,8 +53,8 @@ export interface Duration {
     seconds: number;
 }
 
-const defaultLocaleName = 'en';
-const defaultLocale: Locale = {
+export const defaultLocaleName = 'en';
+export const defaultLocale: Locale = {
     playAria: 'play',
     pauseAria: 'pause',
     replayAria: 'replay',
@@ -64,25 +62,3 @@ const defaultLocale: Locale = {
     formatDuration: durationFormatterForLocale(defaultLocaleName),
     formatRemainingDuration: (duration: string) => `${duration} remaining`
 };
-
-export function getLocale(name: string): Locale {
-    return localesByName[name] ?? defaultLocale;
-}
-
-/**
- * Register a new locale with the given name.
- *
- * The locale's name should preferably be a BCP 47 language identifier, so it can be set as
- * a valid [`lang` attribute](https://developer.mozilla.org/en-US/docs/Web/HTML/Reference/Global_attributes/lang)
- * of the UI.
- *
- * @param name The name of the locale.
- * @param locale The locale.
- */
-export function addLocale(name: string, locale: Partial<Locale>) {
-    localesByName[name] = {
-        ...defaultLocale,
-        ...locale,
-        formatDuration: locale.formatDuration ?? durationFormatterForLocale(name)
-    };
-}
