+++ src/i18n/index.ts
@@ -1,1 +1,2 @@
-export * from './Locale';
+export { type Locale, type Duration, type DurationFormatter } from './Locale';
+export { getLocale, addLocale } from './LocaleRegistry';
